Extract MinIO client functionality to trino-testing-containers module
